Ensure toast still fits on screen after rotation
Currently, if you rotate your device from landscape to portrait with a toast with a long message, it won't fit on screen. This fix will make sure it does, and on top of that it will look exactly as if it was generated in the new position. This should at the very least be an option I feel.